feat(web): full AI fit in the tracking card + reusable JobFitFull
Extract the /fit page's report + live SSE stream into a shared JobFitFull
component and reuse it in both places, so the tracking card's Job Match tab
shows the full analysis (radial gauge, verdict, dimensions, ATS requirements,
strengths/gaps, live stage stepper + thinking) instead of the compact summary
card. The /fit route keeps its masthead and renders <JobFitFull>; the component
seeds from an SSR-cached fit when given one (initial) or fetches the cache on
mount (the card, which has no SSR) — one pure seedFrom() for both paths.

Card tweaks: View job moves up beside the close button (opens in a new tab);
Profile match (deterministic) and Job Match (full fit) are separate tabs; the
tab rail scrolls horizontally on narrow screens (w-max pill in an overflow-x
wrapper, scrollbar hidden) instead of wrapping to two rows.
